jterm: Switch from `perldoc -oman` to `perldoc -oansi`
`-oman` is broken on macOS since whenever they dropped the `-V` option from
`mandoc` (freebsd dropped it in 2015). It was fixed in the [Pod::Perldoc repo][1]
[in 2019][2] but as of today the last non-dev release was in 2017
(3.28) and the most recent dev release (3.28_01) was just a small branch off
of the 2017 codebase and doesn't include the fix. Maybe someday it will get
fixed in a release (and maybe someday someone will have a non-ansi terminal
and will care to change it).

Unfortunately -oansi spews a dozen warnings on the mac so we just hackily
close STDERR before perldoc-ing. Sigh.

[1]: https://github.com/briandfoy/pod-perldoc
[2]: https://github.com/briandfoy/pod-perldoc/commit/3235a0d2b40a43d3897b8a5d0b59ca06b3f6e61c